audit flow/requirements.txt
pandas
pyyaml
requests
beautifulsoup4
pytest
openpyxl